Fix ioctl definition on s390x-musl targets.
ioctl(2) always takes an int as the request, not a ulong, on musl.
I copied a little too closely from the glibc definitions when creating
the original s390x-musl ones.

+++ src/unix/linux_like/linux/musl/b64/s390x.rs
@@ -743,5 +743,5 @@ pub const SYS_epoll_pwait2: ::c_long = 441;
 pub const SYS_mount_setattr: ::c_long = 442;
 
 extern "C" {
-    pub fn ioctl(fd: ::c_int, request: ::c_ulong, ...) -> ::c_int;
+    pub fn ioctl(fd: ::c_int, request: ::c_int, ...) -> ::c_int;
 }
